rename: 모달 관련 폴더와 파일명 변경
